+++ apx_sdk/CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Renamed `CanonicalUrl` type to `CanonicalUri`.
+- Renamed `ApUrl` type to `ApUri`.
+- Renamed `is_ap_url` function to `is_ap_uri`.
+- Renamed `VerificationMethod::ApUrl` variant to `VerificationMethod::ApUri`.
 
 ## [0.18.0] - 2025-10-08